refactor: reorganize discount test files into a dedicated directory
